requirements.txt
Flask
werkzeug
PyPDF2
langchain
python-dotenv
openai
tiktoken
faiss-cpu